+++ pm4py_ucm/__init__.py
@@ -42,4 +42,4 @@ __all__ = [
     "save_vis_ucm",
 ]
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"